+++ src/main/java/com/DailyLife/dto/User.java
@@ -10,15 +10,11 @@ import javax.validation.constraints.Pattern;
 public class User {
 
     private Long uno;
-//    @Pattern(regexp = "^{5,15}(?=.*\\d)(?=.*[a-zA-Z])(?=.*[!@#$%^&+=_]).*$" )
     private String userNickName;
-//    @Pattern(regexp = "^[a-zA-Z0-9]{5,15}$")
     private String userId;
-//    @Pattern(regexp = "^{8,20}(?=.*\\d)(?=.*[a-z])(?=.*[A-Z])(?=.*[!@#$%^&+=_]).*$" )
     private String userPassword;
     private String userPasswordCheck;
     private String userEmail;
-
     private Integer emailAuthor;
 
 }
