+++ src/main/resources/mapper/UserMapper.xml
@@ -32,4 +32,7 @@
         delete from tbl_user where userId = #{userId}
     </delete>
 
+
+
+
 </mapper>
