+++ src/main/resources/templates/coment.html
@@ -24,12 +24,10 @@
 
 <div id="wrapper">
     <div class="fields" style="text-align: center">
-        <a th:href="@{/Index}" class="button primary">HOME</a>
+        <a th:href="@{/index}" class="button primary">HOME</a>
     </div><br>
     <section id="main">
         <!--        <form method="post" action="#" style="position: relative; left: -15%; overflow: auto; height: 200px; width: 500px;" >-->
-
-
 
         <section id="three" class="wrapper style1">
             <header class="major">
@@ -65,7 +63,6 @@
 
                     <div style="width: 70%;">
                         <div class="container">
-
                             <div class="col-4 col-6-medium col-12-small">
                                 <article class="box post" style="width : 100%;">
                                     <article style="width: 70%; height:1200px; margin: 0 auto;">
