+++ src/main/resources/templates/user/findForm.html
@@ -73,7 +73,7 @@
 
 					<br>
 					<ul class="actions special" style="margin-top: 1%;">
-						<li><button type="button" disabled>비밀번호 변경</button></li>
+						<li><button type="button" disabled id="PwChangeButton">비밀번호 변경</button></li>
 					</ul>
 				</div>
 			</div>
